va/requirements.txt
numpy
pandas==1.3.5
scikit-image
opencv-python==4.7.0.72
PyYAML
colorama
tqdm
requests
matplotlib
psutil
gunicorn
flask
scikit-learn
google-cloud-aiplatform
gunicorn